Extract place translation module
